requirements.txt
atlassian-python-api==3.41.11
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
wrapt==1.16.0
